Got it working then broke it again by putting things in functions
Switched to floats; added operater pushing; created basic number entry; began making entry functional which broke calc

+++ include/Node.h
@@ -6,18 +6,18 @@ class Node
 {
     public:
         Node();
-        Node(int val) {value = val;}
+        Node(float val) {value = val;}
         virtual ~Node();
 
         char Value() { return value; }
-        void Setvalue(int val) { value = val; }
+        void Setvalue(float val) { value = val; }
         Node* Next() { return next; }
         void Setnext(Node* val) { next = val; }
 
     protected:
 
     private:
-        int value;
+        float value;
         Node* next;
 };
